Phase 17 Wave B: sensor split + test coverage + silver scale (2.0.0-dev18)
17.4: Split sensor.py (992L) → sensor.py (390L, live UDP sensors) +
sensor_config.py (628L, 400+ config-derived diagnostics). __init__.py's
_apply_config lazy-imports create_config_entities from the new module.

17.7: Grew tests/test_integration.py from 6 tests to 22. Added config-flow
coverage (user happy path + parametrized probe errors + duplicate abort +
hassio confirm + hassio rediscovery URL update + reconfigure happy path +
reconfigure probe error), v1→v2 async_migrate_entry, unknown/malformed WS
payloads, stale vs. live device removal policy, and diagnostics snapshot.
Two flows needed patch("OutbackMate3.start") to avoid aiodns resolver
timer leaks PHACC catches as test failures.

17.12: quality_scale bronze → silver now that 17.1–17.7 have landed.
