+++ outback_mate3_addon/bundled_integration/outback_mate3/manifest.json
@@ -10,7 +10,7 @@
   "hassio": true,
   "iot_class": "local_push",
   "issue_tracker": "https://github.com/weirded/ha-outback-mate3/issues",
-  "quality_scale": "bronze",
+  "quality_scale": "silver",
   "requirements": [],
-  "version": "2.0.0-dev16"
+  "version": "2.0.0-dev18"
 }
