+++ outback_mate3_addon/config.yaml
@@ -1,5 +1,5 @@
 name: "Outback MATE3"
-version: "2.0.0-dev17"
+version: "2.0.0-dev18"
 slug: outback_mate3
 description: >-
   Listens for Outback MATE3 UDP streaming telemetry and relays parsed device
